[Red] non-unit-sized radius test

+++ tddkata/Cylinder.cs
@@ -4,10 +4,10 @@ namespace tddkata
 {
     public class Cylinder : IShape
     {
-        private int radius;
+        private double radius;
         private int length;
 
-        public Cylinder(int radius, int length)
+        public Cylinder(double radius, int length)
         {
             this.radius = radius;
             this.length = length;
